Move types back to .types
